requirements.txt
celery[redis]==4.2.1
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
django-mptt==0.9.1
psycopg2-binary==2.7.5
channels==2.1.3
channels-redis==2.3.0
djangorestframework-jwt==1.11.0
requests==2.19.1
pytest==3.8.0
pytest-timeout==1.3.2
